requirements.txt
tensorflow-cpu>=2.15.0
opencv-python-headless>=4.7.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
streamlit>=1.22.0
Pillow>=9.3.0
